post meeting changes
